new qq-plot component
